requirements.txt
asgiref==3.8.1
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
python-decouple==3.8
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0
django-cors-headers==4.3.1 